+++ src/Management/SearchEngine.php
@@ -43,11 +43,26 @@ class SearchEngine {
   }
 
   /**
-   * Get a list of searchengine's types
+   * Get a list of search engine's user types.
+   *
+   * Some types are internal (their name starts by 'df_') and this method does
+   * not return them.
    *
    * @return array list of types
    */
   public function getTypes() {
+    $lambda = function($name) {
+      return strpos($name, 'df_') !== 0;
+    };
+    return array_filter($this->getAllTypes(), $lambda);
+  }
+
+  /**
+   * Get a list of all search engine's types (both user and internal types).
+   *
+   * @return array list of types
+   */
+  public function getAllTypes() {
     $result = $this->client->managementApiCall('GET', "{$this->hashid}/types");
     return $result['response'];
   }