+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [5.8.0] - 2018-10-23
 ### Added
   - `getItems()` method in order to deprecate the old `items()` method.
   - Deprecation message for `getDatatypes()` method in favor of the `getTypes()` method.